Conditional delete of flavor in test tearDown
This patch fixes a bug where the test tearDown was deleting flavor
that was not created by the test.

+++ tests/api/services/test_services.py
@@ -319,7 +319,8 @@ class TestDeleteService(base.TestBase):
 
     def tearDown(self):
         self.client.delete_service(service_name=self.service_name)
-        self.client.delete_flavor(flavor_id=self.flavor_id)
+        if self.test_config.generate_flavors:
+            self.client.delete_flavor(flavor_id=self.flavor_id)
         super(TestDeleteService, self).tearDown()
